Route users based on auth state
