+++ navigation/AuthStack.js
@@ -1,0 +1,18 @@
+// navigation/AuthStack.js
+import React from 'react';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+import LoginScreen from '../screens/LoginScreen';
+import EmailLoginScreen from '../screens/EmailLoginScreen';
+import SignUpScreen from '../screens/SignUpScreen';
+
+const Stack = createNativeStackNavigator();
+
+export default function AuthStack() {
+  return (
+    <Stack.Navigator screenOptions={{ headerShown: false }}>
+      <Stack.Screen name="Login" component={LoginScreen} />
+      <Stack.Screen name="EmailLogin" component={EmailLoginScreen} />
+      <Stack.Screen name="Signup" component={SignUpScreen} />
+    </Stack.Navigator>
+  );
+}
