+++ navigation/OnboardingStack.js
@@ -1,0 +1,16 @@
+// navigation/OnboardingStack.js
+import React from 'react';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+import OnboardingScreen from '../screens/OnboardingScreen';
+import AppStack from './AppStack';
+
+const Stack = createNativeStackNavigator();
+
+export default function OnboardingStack() {
+  return (
+    <Stack.Navigator screenOptions={{ headerShown: false }}>
+      <Stack.Screen name="Onboarding" component={OnboardingScreen} />
+      <Stack.Screen name="Main" component={AppStack} />
+    </Stack.Navigator>
+  );
+}
