+++ navigation/RootNavigator.js
@@ -1,31 +1,15 @@
 // navigation/RootNavigator.js
 import React, { useEffect, useState } from 'react';
-import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import { ActivityIndicator, StatusBar, Text, View } from 'react-native';
 import * as Linking from 'expo-linking';
 import { useUser } from '../contexts/UserContext';
 
 import SplashScreen from '../screens/SplashScreen';
-import LoginScreen from '../screens/LoginScreen';
-import EmailLoginScreen from '../screens/EmailLoginScreen';
-import SignUpScreen from '../screens/SignUpScreen';
-import ProfileScreen from '../screens/ProfileScreen';
-import MainTabs from './MainTabs';
-import ChatScreen from '../screens/ChatScreen';
-import NotificationsScreen from '../screens/NotificationsScreen';
-import EditProfileScreen from '../screens/EditProfileScreen';
-import GameInviteScreen from '../screens/GameInviteScreen';
-import GameLobbyScreen from '../screens/GameLobbyScreen';
-import CommunityScreen from '../screens/CommunityScreen';
-import EventChatScreen from '../screens/EventChatScreen';
-import PremiumScreen from '../screens/PremiumScreen';
-import PremiumPaywallScreen from '../screens/PremiumPaywallScreen';
-import StatsScreen from '../screens/StatsScreen';
-import OnboardingScreen from '../screens/OnboardingScreen';
+import AuthStack from './AuthStack';
+import AppStack from './AppStack';
+import OnboardingStack from './OnboardingStack';
 
 
-
-const Stack = createNativeStackNavigator();
 const splashDuration = 2000;
 
 export default function RootNavigator() {
@@ -54,36 +38,13 @@ export default function RootNavigator() {
 
   const onboarded = user?.onboardingComplete;
 
-  return (
-    <Stack.Navigator screenOptions={{ headerShown: false }}>
-      {!user ? (
-        <>
-          <Stack.Screen name="Login" component={LoginScreen} />
-          <Stack.Screen name="EmailLogin" component={EmailLoginScreen} />
-          <Stack.Screen name="Signup" component={SignUpScreen} />
-          <Stack.Screen name="Main" component={MainTabs} />
-        </>
-      ) : onboarded ? (
-        <>
-          <Stack.Screen name="Main" component={MainTabs} />
-          <Stack.Screen name="Profile" component={ProfileScreen} />
-          <Stack.Screen name="Chat" component={ChatScreen} />
-          <Stack.Screen name="Notifications" component={NotificationsScreen} />
-          <Stack.Screen name="EditProfile" component={EditProfileScreen} />
-          <Stack.Screen name="GameInvite" component={GameInviteScreen} />
-          <Stack.Screen name="GameLobby" component={GameLobbyScreen} />
-          <Stack.Screen name="Community" component={CommunityScreen} />
-          <Stack.Screen name="EventChat" component={EventChatScreen} />
-          <Stack.Screen name="Premium" component={PremiumScreen} />
-          <Stack.Screen name="PremiumPaywall" component={PremiumPaywallScreen} />
-          <Stack.Screen name="Stats" component={StatsScreen} />
-        </>
-      ) : (
-        <>
-          <Stack.Screen name="Onboarding" component={OnboardingScreen} />
-          <Stack.Screen name="Main" component={MainTabs} />
-        </>
-      )}
-    </Stack.Navigator>
-  );
+  if (!user) {
+    return <AuthStack />;
+  }
+
+  if (!onboarded) {
+    return <OnboardingStack />;
+  }
+
+  return <AppStack />;
 }
